+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 ## 1.0.2
 
 - Changed minimal SDK version to `2.12.0` to indicate the full null-safety support.
+
+## 1.1.0
+
+- Fixed the signature of the methods, so they can take JSON in any cases.
+- Added support for JSON arrays with `processAsList` method.
+- Added the `extract` option to both `process` and `processAsList` methods.